+++ src/app.module.ts
@@ -10,6 +10,7 @@ import { RedisModule, REDIS_CLIENT } from './common/redis/redis.module.js';
 import { RedisThrottlerStorage } from './common/throttling/redis-throttler.storage.js';
 import { CacheModule } from './common/cache/cache.module.js';
 import { StorageModule } from './common/storage/storage.module.js';
+import { MailModule } from './common/mail/mail.module.js';
 import { HttpExceptionFilter } from './common/filters/http-exception.filter.js';
 import { CommonModule } from './common/common.module.js';
 import { PrismaModule } from './prisma/prisma.module.js';
@@ -97,6 +98,7 @@ import { RedisClient } from './common/redis/redis-client.interface.js';
     RedisModule,
     CacheModule,
     StorageModule,
+    MailModule,
     HealthModule,
     PrismaModule,
     AuthModule,